src/containers/modbus_nats_kv/requirements.txt
nats-py~=2.2
pymodbus~=3.0.2
